Page-Agile/Workspace-reservation: move data fetching store && Select Component styling
