+++ src/screens/agileOffice/screens/AgileOfficeWorkspaceReservation/components/index.tsx
@@ -1,0 +1,1 @@
+export { default as Select } from "./Select";
