+++ src/screens/agileOffice/screens/AgileOfficeWorkspaceReservation/types.ts
@@ -1,0 +1,4 @@
+export type TimeType = {
+  from: Date;
+  to: Date;
+};
