+++ src/types.ts
@@ -1,0 +1,4 @@
+export interface Option {
+  key: string;
+  label: string;
+}
